refactor: dataset refresh button
